# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: ELOJIM API
  description: API oficial del proyecto Fundación ELOJIM
  version: "1.0.0"
  contact:
    name: Soporte ELOJIM
    email: soporte@elojim.org
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    RegistroBase:
      type: object
      properties:
        nombreCompleto:
          type: string
          example: "Laura Gómez"
        fechaNacimiento:
          type: string
          format: date
        edad:
          type: integer
        comuna:
          type: string
        estratoSocial:
          type: string
        grupoEtnico:
          type: string
        telefonoContacto:
          type: string
        correoElectronico:
          type: string
        direccion:
          type: string
        motivacion:
          type: string
        expectativas:
          type: string
        aceptaTerminos:
          type: boolean
      required:
        - nombreCompleto
        - fechaNacimiento
        - comuna
        - estratoSocial
        - telefonoContacto
        - correoElectronico
        - aceptaTerminos
...
    RegistroSoftwareFactory:
      allOf:
        - $ref: "#/components/schemas/RegistroBase"
        - type: object
          properties:
            tipoDocumento:
              type: string
            numeroDocumento:
              type: string
            modalidadVinculacion:
              type: string
            institucionEducativa:
              type: string
            programaAcademico:
              type: string
            semestreNivel:
              type: string
            tiempoDisponible:
              type: string
            tecnologias:
              type: array
              items:
                type: string
            proyectosRealizados:
              type: string
            areasInteres:
              type: array
              items:
                type: string
            experienciaAgile:
              type: string
          example:
            nombreCompleto: "Alejandro Martínez"
            modalidadVinculacion: "Pasantía"
            institucionEducativa: "Universidad de Antioquia"
            tecnologias: ["Desarrollo web", "Backend (Node.js)"]
            aceptaTerminos: true
    RegistroTallerSteam:
      allOf:
        - $ref: "#/components/schemas/RegistroBase"
        - type: object
          properties:
            institucionEducativa:
              type: string
            cursoGrado:
              type: string
            nombreAcudiente:
              type: string
            relacionNino:
              type: string
            participacionPrevia:
              type: boolean
            actividadesInteres:
              type: array
              items:
                type: string
            otrasActividades:
              type: string
            disponibilidad:
              type: string
            accesoComputadora:
              type: boolean
            accesoInternet:
              type: boolean
            expectativa:
              type: string
          example:
            nombreCompleto: "Camila Rodríguez"
            actividadesInteres: ["Programación básica", "Ciencia y experimentos"]
            accesoComputadora: true
            expectativa: "Crear mi primer videojuego"
            aceptaTerminos: true
    RegistroVoluntariado:
      allOf:
        - $ref: "#/components/schemas/RegistroBase"
        - type: object
          properties:
            tipoDocumento:
              type: string
            numeroDocumento:
              type: string
            nivelEducativo:
              type: string
            profesionOcupacion:
              type: string
            disponibilidadTipo:
              type: string
            diasEspecificos:
              type: string
            horasDisponibles:
              type: integer
            areasInteres:
              type: array
              items:
                type: string
            otrasAreas:
              type: string
            habilidades:
              type: string
            fundacion:
              type: string
            funcion:
              type: string
            tiempo:
              type: string
            referencia1Nombre:
              type: string
            referencia1Telefono:
              type: string
            referencia2Nombre:
              type: string
            referencia2Telefono:
              type: string
          example:
            nombreCompleto: "Laura Martínez"
            profesionOcupacion: "Psicóloga"
            areasInteres: ["Educación", "Medio ambiente"]
            habilidades: "Escucha activa, liderazgo"
            aceptaTerminos: true
...
security:
  - cookieAuth: []
paths:
...
  /api/registro/software-factory:
    get:
      tags: [Registro]
      summary: Listar registros de software factory
      responses:
        "200":
          description: Lista de registros
    post:
      tags: [Registro]
      summary: Crear registro en Software Factory
      description: Crear un nuevo registro de participantes en Software Factory.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistroSoftwareFactory"
      responses:
        "201":
          description: Registro creado exitosamente
        "400":
          description: Error en validación o datos incompletos
        "500":
          description: Error interno
    delete:
      tags: [Registro]
      summary: Eliminar un registro de software factory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 16
      responses:
        "200":
          description: Registro eliminado exitosamente
        "404":
          description: Registro no encontrado
        "500":
          description: Error interno
  /api/registro/taller-steam:
    get:
      tags: [Registro]
      summary: Listar registros de taller steam
      responses:
        "200":
          description: Lista de registros
    post:
      tags: [Registro]
      summary: Crear registro para taller STEAM
      description: Crear un nuevo participante en el programa Taller Steam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistroTallerSteam"
      responses:
        "201":
          description: Registro creado exitosamente
        "400":
          description: Error en validación o datos incompletos
        "500":
          description: Error interno
    delete:
      tags: [Registro]
      summary: Eliminar un registro de taller steam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 16
      responses:
        "200":
          description: Registro eliminado exitosamente
        "404":
          description: Registro no encontrado
        "500":
          description: Error interno
  /api/registro/voluntariado:
    get:
      tags: [Registro]
      summary: Listar registros de voluntariado
      responses:
        "200":
          description: Lista de registros
    post:
      tags: [Registro]
      summary: Crear registro para Voluntariado
      description: Crear un nuevo participante para el programa Voluntariado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistroVoluntariado"
      responses:
        "201":
          description: Registro creado exitosamente
        "400":
          description: Error en validación o datos incompletos
        "500":
          description: Error interno
    delete:
      tags: [Registro]
      summary: Eliminar un registro de voluntariado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 16
      responses:
        "200":
          description: Registro eliminado exitosamente
        "404":
          description: Registro no encontrado
        "500":
          description: Error interno      
...